[skip-ci] Can construct arrays on GPU using the Delayed worker
